agents/requirements.txt
# Core dependencies
langchain-openai>=0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0

# IPFS
ipfshttpclient>=0.8.0a1
requests>=2.31.0

# Solana dependencies
anchorpy>=0.18.0
solders>=0.18.0
solana>=0.30.0
base58>=2.1.1

